Netlib

+++ NetworkLibrary/src/Core/Address.cpp
@@ -4,7 +4,10 @@
 
 namespace NetLib
 {
-	Address::Address(const std::string& ip, unsigned int port)
+	//Evaluate if supporting address creation from hostname such as hello.com. In order to do that you will need to use getaddrinfo
+	//and create a DNS petition which takes wayyyy longer than using just a normal IP. If so, do it in a separate thread.
+
+	Address::Address(const std::string& ip, uint16_t port)
 	{
 		_addressInfo.sin_family = AF_INET;
 		_addressInfo.sin_port = htons(port);
